Improved templates, add account creation and cache
Added a unit test, will add more later.
First pass at creating HTML components and store them in a subdirectory.
This way, template usage will be cleaner.

Added a AccountData cache in the ApiConfig, so that it doesn't hit the
database that often. Still have to check if there are other places where
it needs to refresh the cache, of if there are any better caching
options, like breaking entities apart.

Added a bit of CSS, just to darken things.